classifier/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
httpx==0.27.2
pyyaml==6.0.2
